log: Fix compiler errors
Newer compiler versions do not like passing single arguments to
printf/printk where it does not know if the single string contains
a format or not.  The simple fix is to always pass in "%s" as the
first arg in these cases.  So modify the macros to add check on
the number of args passed to __VA_ARGS__ and pass in "%s" when
there is only one arg.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>