testcases/ddt/scripts/eth/netlib.sh: Add helper to dump ALE entries
Add helper function to dump ALE entries for a given interface, to a
given file, using the switch-config tool. The output is pre-processed by
removing numbers and sorting the entries, to ensure that the order in
which the entries are added to the ALE doesn't affect the output.

This can be used to compare the ALE entries before and after
suspend/resume for example.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>