testcases/ddt/scripts/eth/netlib.sh: Fix ping test
The existing ping test attempts to ping the DHCP server and relies on
journalctl logging the DHCP server's IP. However, depending on the log
level, it is possible that the DHCP server's IP is not logged.

Fix this by using a new approach with two levels of fallback options.
The new approach involves verifying ping in the following sequence,
until at least one of them works:
1. Try pinging the IPERF server if the host exports the IP address
   to the DUT using the IPERFHOST variable.
2. Try pingint the DHCP server if journalctl log captures the IP address
   of the DHCP server.
3. Try broadcast ping which verifies whether any of the devices
   listening to the broadcast reply to the ping request.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>