testcases/ddt/scripts/eth/icssg_run_basic_tests.sh: Bring ICSSG interfaces up if down.
ICSSG interfaces are not up by default in NFS boot. This commits adds
support for bringing ICSSG interfaces up if they are not up by default.
Before running any icssg related testcase, it'll check if interface is
down and then try to bring it up. If interface is already up, it will
run the testcase.

Signed-off-by: MD Danish Anwar <danishanwar@ti.com>