Crypto: fix the logic to get the interrupt count for sa2ul
Earlier logic used to consider the channel number while doing grep
but this does not work well with am62 as it has 2 sets of dma-controller
and hence 2 set for each channel number. The current logic is flawed
since it consider only top most interrupt count entry for each channel
number. Fix this by using "Level" interrupts as filter for grep. And then
adding the interrupt count for all Level interrupts for dma-controller.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>