common.sh: swapfile destroy only called if create was called
Swapfile destroy command will now only destroy the swapfile if we
explicitly created it by using a trap. This suppresses the warning that
users see when die attempts to clean a non-existent swapfile and also
avoids the cleanup command replacing the exit code DDT uses to verify
the test passes.

Signed-off-by: Randolph Sapp <rs@ti.com>