testcases/ddt/scripts/gpio/gpio.sh: Use libgpiod library instead of gpio-sysfs
/sys/class/gpio interface is deprecated, hence the testcase is re-written
using libgpiod library instead of the gpio-sysfs.

GPIOINFO is used to extract the information regarding the unused output
gpio pins. GPIOSET is then used to set the value of a gpio pin for a certain
time. During this period of time, GPIOINFO is used to check the correctness
of GPIOSET.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>