piglit: add a bilinear / trilinear filtering test
Add a test that runs all linear tests and checks the output. The test
will need to be analyzed further on failure to see what form of
filtering was used.

Piglit has a series of linear tests that utilize bilinear and trilinear
comparison tools depending on the test. Unfortunately they wrap CTS and
they just group all of these under the linear keyword.

Signed-off-by: Randolph Sapp <rs@ti.com>