ddt: scripts: setupcamera.sh: Update formats on all subdevs
With ti-linux kernel 6.6 formats on all v4l2 subdevs in the media
pipeline should be set, as the upstream v4l2 APIs require each subdev
driver to validate links to the connected subdevs.

Make sure we update format on cdns-csi2rx bridge and ticsi2rx shim
subdevs for IMX219 and OV5640 in the setup script.

Signed-off-by: Jai Luthra <j-luthra@ti.com>