All: Fix the cpuloadgen command by using new wrapper function
With latest cpuloadgen (v0.94) integrated in meta-arago, there is an
update to usage of the utility. Update scripts to the new usage.

Earlier cpuloadgen (v0.93) only allowed loading upto 2 cpu cores. With this
update, we can specify loading for all available cpus. This will also be
a change in behavior of the testcase

functions.sh:
- Add new function cpu_load_nproc() as a blocking wrapper for cpuloadgen
- Update cpu_load_random() function to new usage of cpuloadgen

Affects:
- runtest/ddt/realtime_cpuload_*
- testcases/ddt/scripts/powermgr/super-pm-tests.sh

Signed-off-by: Aniket Limaye <a-limaye@ti.com>