blk_device_common.sh: Create third partition
The blk_device_common script does not create any partition if any
partition is found. This may be an issue if the existing partitions
are populated with rootfs or bootloader files. Then the MMC performance
test cases will fail to execute.

Therefore, create a third partition if we find only two partitions.

Signed-off-by: Judith Mendez <jm@ti.com>