testcases: ddt: scripts: ipc: Fix AM64x rproc support for PRU Echo test
The script was not allowing the AM64x to load firmware, start, or stop any
remote proc for the reason that K3 devices do not support toggling the
remote procs. This is out of date and K3 devices do support toggling
remote procs. The IPC_S_FUNC_PRU_ECHO test will now run and pass for the
AM64x.