ltp-ddt: runtest: ddt: Add LPM cases for JPEG Encoder
IMG JPEG Encoder supports LPM. This has been validated using manual testing
in the past but want to offload some of the manual efforts. So in this patch
are cases for the 3 portions of LPM that need to be validated each release:
runtime PM, system suspend, and active suspend/resume.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>