ddt: scripts: piglit_execution.sh
Testcases would incorrectly show up as PASS from the point of view of ltp, when
in fact they had FAILED. This was caused due to a change in Piglit, where failed
testcases are no longer labeled as "failed" but instead "fail". So, update the
search parameter within piglit_execution to adapt to that change.

Expanded search functionality for what constitutes a failure for a test. The
status "skip" should be considered a failure, as the test-lists that piglit
executes have been curated to not have any tests skipped. So if any skips
arrise, then it is unexpected behavior and a failure.

Singed-off-by: Antonios Christidis <a-christidis@ti.com>