ddt: scripts: piglit_execution.sh: test script improvements
Weston is a dependency for piglit tests to run, if it is not present all
testcases will fail.  The test script is now checking is weston is running. If
not it, tries to start it before causing the test to fail.

Also, added the kmsprint command. This will provide us info about the display
connected which maybe useful for future debugging of failures.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>