testcases/ddt/scripts/pci/pcielib: relax pass criteria for link width test
The Link Width test script looks for an exact match. However, for the
sake of simplicity in terms of device setup, treat a higher link width
as a pass. This shall reduce the number of distinct device setups
required, wherein just a single setup connecting the PCIe connectors
corresponding to the maximum Link Width supported by the device is
sufficient.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>