runtest: ddt: tidec_decode: Fix Failing Test Cases
The cases being changed in this commit are failing for 2 reasons. First,
the new verification method trying to be implemented is not executing
correctly leading to false failures. Second, the checksums in the videos
are outdated. This commit removes the new verification method for those
test cases and adds checksum based verification instead. Once this new
verification issue has been verified, it will be added back.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>