ddt: hwspinlock: Switch log level for modprobe
In RT-kernel, modprobe is non-blocking so we get the console back
before spitting out all dmesg logs, this fails the test only on
RT kernel since the rest of the test is not carried out. Fix by
changing the log level and not flooding the console.

The one side affect of this is that debug messages from modprobe
will not show anymore.

Signed-off-by: Judith Mendez <jm@ti.com>