ddt: scripts: Fix sa2ul interrupt counter skipping the last irq num
The commit fbb7b50a7 ("Crypto: fix the logic to get the interrupt count
for sa2ul") changed logic for interrupt counting across different irqs.
Earlier we were iterating over dma-controller channels which are named
chan0, chan1, ... They are zero indexed. But the above commit changed
it to get the i-th channel using head and tail commands.

The problem here is that head needs the number of lines starting from 1
and not 0. But the loop was not changed there. As a result, in the first
iteration, i = 0 would cause `head -n 0` to return no lines at all. And
the last iteration will have i = <num_channels> - 1 and we only count
till num_channels - 1 channels. Completely missing the last channel.
This commit fixes the loop here.

Fixes: fbb7b50a7 ("Crypto: fix the logic to get the interrupt count for sa2ul")
Signed-off-by: T Pratham <t-pratham@ti.com>