functions.sh: Add option in suspend to detach the console
Detaching the console stops prints from happening on the console which
increases the speed at which resume happens. This can be helpful to do
when trying to measure the resume latency of Linux since prints add
unnecessary time to resume. Detaching the console allows for more
accurate measurement of Linux resume latency.

Signed-off-by: Kendall Willis <k-willis@ti.com>