functions.sh: Remove redundant case statement
The case statement sets two parameters values based off of machine, but
it does this for ALL machines. Therefore, this case statement is not
needed since the logic executes regardless of what machine is being
tested. Fix this by removing the case statement and moving the logic
outside of it.

Signed-off-by: Kendall Willis <k-willis@ti.com>